Mv dbapi base class to proxy.py
